+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.8.3
 
 * Attempting to resolve entry file rather than checking the file's existence.
+* Better priorities when solving webpack config.
 * Fixing update logs.
 
 ## 0.8.2